requirements.txt
asgiref==3.5.2
asttokens==2.0.8
attrs==22.1.0
autopep8==1.7.0
backcall==0.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
colorama==0.4.5
cryptography==37.0.2
decorator==5.1.1
Django==4.0.4
django-crontab==0.7.1
django-jazzmin==2.5.0
djangorestframework==3.13.1
executing==1.0.0
idna==3.3
iniconfig==1.1.1
ipdb==0.13.9
ipython==8.5.0
jedi==0.18.1
matplotlib-inline==0.1.6
numpy==1.22.4
packaging==21.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.1.1
pluggy==1.0.0
prompt-toolkit==3.0.31
pure-eval==0.2.2
py==1.11.0
pyasn1==0.4.8
pycodestyle==2.9.1
pycparser==2.21
Pygments==2.13.0
pyparsing==3.0.9
pytest==7.1.3
pytest-django==4.5.2
pytz==2022.1
requests==2.27.1
rsa==4.8
six==1.16.0
sqlparse==0.4.2
stack-data==0.5.0
toml==0.10.2
tomli==2.0.1
traitlets==5.3.0
tzdata==2022.1
urllib3==1.26.9
wcwidth==0.2.5
